need to fix repo again

+++ src/transformers/models/minimax_m3_vl/modeling_minimax_m3_vl.py
@@ -1047,7 +1047,7 @@ class MiniMaxM3VLVisionRotaryEmbedding(nn.Module):
 
         cos = self.recomposition_frequencies(cos)
         sin = self.recomposition_frequencies(sin)
-        return cos, sin
+        return cos.to(x.dtype), sin.to(x.dtype)
 
     def recomposition_frequencies(self, freq):
         """
